+++ docs/_pages/releases.md
@@ -10,6 +10,7 @@ sidebar:
 ## Unreleased
 
 ### What's new
+* Add `ContainInExplicitOrder` and `NotContainInExplicitOrder` assertions to check if a collection contains or not items in a specific order and to be consecutive. - [#](https://github.com/fluentassertions/fluentassertions/pull/)
 
 ### Fixes
 * Fix `For`/`Exclude` not excluding properties in objects in a collection - [#1953](https://github.com/fluentassertions/fluentassertions/pull/1953)